refactor: rename operatorDefaults to operator Preset
